Preserve @ symbols in vcloud_director usernames
It's valid to have an email address as a vCloud Director username which
would result in a `vcloud_director_username` string that looks like:

    user@example.com@organisation

The previous code incorrectly stripped any @ symbols from the username which
would turn the previous example into:

    fog_credential_user = 'userexample.com'
    fog_credential_org = 'organisation'

This meant that check_session_matches_credentials() would always raise an
error for such usernames. Prevent this from happening by restoring the @
separators when we join the array back together (minus the org).

Fixes an error reported in gds-operations/vcloud-core#139