[Openstack|Compute] fix randomly failing spec
The security group failure tests have been failing randomly. This happens
because the random id assigned by the create security group mock can collide
with the id (0) of the default security group. The mock data stores the keys
as strings (except) for the default one, meaning that you end up with 2
security groups with id 0, one stored under the key 0 and the other
under the key '0'